+++ noodles-sam/src/alignment/record/flags.rs
@@ -5,7 +5,10 @@ bitflags::bitflags! {
         /// Read is segmented (`0x01`).
         const SEGMENTED = 0x01;
         /// Each segment in the read is properly aligned (`0x02`).
+        #[deprecated(since = "0.54.0", note = "Use `PROPERLY_SEGMENTED` instead.")]
         const PROPERLY_ALIGNED = 0x02;
+        /// Each segment in the read is properly aligned (`0x02`).
+        const PROPERLY_SEGMENTED = 0x02;
         /// Read is unmapped (`Ox04`).
         const UNMAPPED = 0x04;
         /// The mate is unmapped (`0x08`).
@@ -52,8 +55,23 @@ impl Flags {
     /// assert!(Flags::PROPERLY_ALIGNED.is_properly_aligned());
     /// assert!(!Flags::UNMAPPED.is_properly_aligned());
     /// ```
+    #[allow(deprecated)]
+    #[deprecated(since = "0.54.0", note = "Use `Flags::is_properly_segmented` instead.")]
     pub fn is_properly_aligned(self) -> bool {
         self.contains(Self::PROPERLY_ALIGNED)
+    }
+
+    /// Returns whether the `PROPERLY_SEGMENTED` flag is set.
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// use noodles_sam::alignment::record::Flags;
+    /// assert!(Flags::PROPERLY_SEGMENTED.is_properly_segmented());
+    /// assert!(!Flags::UNMAPPED.is_properly_segmented());
+    /// ```
+    pub fn is_properly_segmented(self) -> bool {
+        self.contains(Self::PROPERLY_SEGMENTED)
     }
 
     /// Returns whether the `UNMAPPED` flag is set.
@@ -199,6 +217,7 @@ impl From<Flags> for u16 {
     }
 }
 
+#[allow(deprecated)]
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -208,9 +227,9 @@ mod tests {
         let flags = Flags::default();
 
         assert!(flags.is_empty());
-
         assert!(!flags.is_segmented());
         assert!(!flags.is_properly_aligned());
+        assert!(!flags.is_properly_segmented());
         assert!(!flags.is_unmapped());
         assert!(!flags.is_mate_unmapped());
         assert!(!flags.is_reverse_complemented());
@@ -227,6 +246,7 @@ mod tests {
     fn test_contains() {
         assert!(Flags::SEGMENTED.is_segmented());
         assert!(Flags::PROPERLY_ALIGNED.is_properly_aligned());
+        assert!(Flags::PROPERLY_SEGMENTED.is_properly_segmented());
         assert!(Flags::UNMAPPED.is_unmapped());
         assert!(Flags::MATE_UNMAPPED.is_mate_unmapped());
         assert!(Flags::REVERSE_COMPLEMENTED.is_reverse_complemented());